Update angular-soundmanager2.js
line 6160: 
I had problems when trying to play a song that doesn't exist (streaming it from my machine and not it from a url).

line 6256 / 6279:
When starting the player, if the user click on next or previous song, the player crashes because there is no current track by then and we try to get the previous or next track key. This prevents not desired results if the user click on prev or next before playing the first song (loading it).

line 6342.
adjustVolumeSlider is a new factory function to allow having volume not only with buttons but with a slider. Note that in this case we can change the volume with a range of 1 by 1 and not 10 by 10.

line 6632:
volumeBar is a new directive inspired on the seekBar, in this case the duration is 100 (max volume), we get the location of the bar where the user clicked and then we send that to our function adjustVolumeSlider created above. Allowing now to use now this directive in our html and have a volume control type slider.

Example:
 <div class="barBase" volume-bar>
                    <div class="barLoad" ng-style="{width: ( volume + '%')}"></div>
</div>